pyproject.toml
[project]
name = "langchain-agentskills"
version = "0.1.0"
description = "agentskills.io support for LangChain and LangGraph"
readme = "README.md"
authors = [
    { name = "AnhQuanTrl", email = "6561860+AnhQuanTrl@users.noreply.github.com" }
]
requires-python = ">=3.13"
dependencies = [
    "langchain-core>=0.3.0",
    "langchain[example,openai]>=1.2.10",
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
examples = [
    "langchain>=0.3.0",
]

[build-system]
requires = ["uv_build>=0.8.0,<0.9"]
build-backend = "uv_build"
